Update module dependencies in test and main code
Module names have been updated in the main and test code's module-info.java files. The previous module "pro.verron.opcstamper" is replaced with "pro.verron.docxstamper" to reflect the changes in the code and its dependencies. This will ensure that all dependencies are correctly loaded and used by both main and test code.