Add detailed description to ITypeResolver interface
A comprehensive description has been added to the ITypeResolver interface to clarify its role of mapping a specific Java object to an object of the DOCX4J API. Furthermore, it has been marked as deprecated with details regarding its newer, more efficient, and versatile replacement — ObjectResolver. This is part of ongoing efforts to simplify and improve the internal workings of the docx-stamper project.