refactor: change ExpressionResolver constructor to take ExpressionParser
Changed the constructor of ExpressionResolver to take an instance of ExpressionParser instead of SpelParserConfiguration. The ExpressionParser object is now created in DocxStamper and passed to the ExpressionResolver construction. Removed unused SpelParserConfiguration import in ExpressionResolver. This change makes ExpressionResolver class more flexible to use either SpelExpressionParser or any other implementation of ExpressionParser.